requirements.txt
alembic==1.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.0
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytesseract==0.3.9
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
slack-bolt==1.14.0
slack-sdk==3.17.2
SQLAlchemy==1.4.39
urllib3==1.26.9
Werkzeug==2.1.2
boto3==1.24.57
botocore==1.27.57